Make some functionality from GrailsFactoriesLoader available via grails-bootstrap
