+++ grails-core/src/main/groovy/org/grails/core/io/support/GrailsFactoriesLoader.groovy
@@ -17,13 +17,8 @@
 package org.grails.core.io.support
 
 import groovy.transform.CompileStatic
-
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.ConcurrentMap
-
+import org.grails.io.support.FactoriesLoaderSupport
 import org.springframework.core.OrderComparator
-import org.springframework.core.io.UrlResource
-import org.springframework.core.io.support.PropertiesLoaderUtils
 import org.springframework.util.Assert
 import org.springframework.util.ClassUtils
 
@@ -36,14 +31,9 @@ import org.springframework.util.ClassUtils
  * @author Graeme Rocher
  */
 @CompileStatic
-class GrailsFactoriesLoader {
+class GrailsFactoriesLoader extends FactoriesLoaderSupport {
 
 
-    /** The location to look for the factories. Can be present in multiple JAR files. */
-    static final String FACTORIES_RESOURCE_LOCATION = "META-INF/grails.factories"
-
-    private static ConcurrentMap<Integer, Map<String,String[]>> loadedPropertiesForClassLoader = new ConcurrentHashMap<Integer, Map<String,String[]>>()
-    
     private static final Object[] NO_ARGUMENTS = [] as Object[]
     
     
@@ -93,41 +83,6 @@ class GrailsFactoriesLoader {
         return result
     }
 
-    static String[] loadFactoryNames(Class<?> factoryClass, ClassLoader classLoader = GrailsFactoriesLoader.class.classLoader) {
-        String factoryClassName = factoryClass.getName()
-        try {
-            Map<String,String[]> loadedProperties = loadedPropertiesForClassLoader.get(System.identityHashCode(classLoader))
-            if( loadedProperties  == null) {
-                Set<String> allKeys = [] as Set
-                def urls = classLoader.getResources(FACTORIES_RESOURCE_LOCATION);
-                Collection<Properties> allProperties = []
-                urls.each { URL url ->
-                    def properties = PropertiesLoaderUtils.loadProperties(new UrlResource(url))
-                    allProperties << properties
-                    allKeys.addAll((Set<String>)properties.keySet())
-                }
-                Map<String,String[]> mergedFactoryNames=[:]
-                for (String propertyName : allKeys) {
-                    Set<String> result = [] as Set
-                    for (Properties props : allProperties) {
-                        String factoryClassNames = props.getProperty(propertyName)
-                        if (factoryClassNames) {
-                            result.addAll factoryClassNames.split(',').toList()
-                        }
-                    }
-                    mergedFactoryNames.put propertyName, result as String[]
-                }
-                loadedProperties = loadedPropertiesForClassLoader.putIfAbsent(System.identityHashCode(classLoader), mergedFactoryNames)
-                if (loadedProperties == null) {
-                    loadedProperties = mergedFactoryNames
-                }
-            }
-            return loadedProperties.get(factoryClassName)
-        }
-        catch (IOException ex) {
-            throw new IllegalArgumentException("Unable to load [$factoryClassName] factories from location [$FACTORIES_RESOURCE_LOCATION]", ex);
-        }
-    }
 
     private static <T> Class<? extends T> loadFactoryClass(String instanceClassName, Class<T> factoryClass, ClassLoader classLoader) {
         try {
